requirements.txt
discord.py
wavelink
aiohttp
python-dotenv
pymongo
